Debug android_instrumentation_apk.
Summary:
Temporary workaround to fix `android_instrumentation_apk`. Apparently, we
were using the APK-under-test's pre-dexed files in the `classes.dex`
for the instrumentation APK, which is completely wrong.

This only became an issue recently as a result of
introducing support for dex-merging for multi-dex apps, which made
`apkUnderTest.getPreDexDeps()` non-empty, which caused `AndroidInstrumentationApk`
to take a different path through the `addDxAndApkBuilderSteps()` method in
its superclass.