Introduce Closure Templates.
Summary:
I decided that I should not write any more Java code that generates
HTML, as using a template language makes a lot more sense.
(We should also consider rewriting the Python code for `buck project`
in Java now that we can template XML with Closure Templates, too.)

I spent hours today trying to serve JSPs from embedded Jetty, but was
unsuccessful. The docs here are incomplete:
https://wiki.eclipse.org/Jetty/Tutorial/Embedding_Jetty

I also tried building Mustache (https://github.com/spullara/mustache.java),
but that pulled in code for Scala and JRuby, so that seemed like it
would result in adding a bunch of stuff that has no business being on
Buck's classpath. Apache's Veclocity was a similar situation.
This pulls in a version of Closure Templates with its
Guava and Java annotation classes stripped out of it so
they do not conflict with the ones we already have on
our classpath.

This will be used for upcoming work on the buckd HTTP server.

Test Plan: Manual testing in follow-up diff.