Clean up the sh_binary script and symlink the scriptToRun.
Summary:
Previously we were running scriptToRun out of the main repo, and
scripts could (and did!) read `dirname $0` to access files in the repo without
declaring them as dependencies. To prevent this, symlink scriptToRun into our
temp dir and execute it from there. Also, make the generated script iterate
over an array of paths instead of duplicating a lot of text, to make the whole
thing more readable.