Use Closure Templates to generate the HTML for the buckd web server.
Summary:
Use a template to create a standard header and footer for all of our pages.
Create a page that shows the list of traces in the `buck-out/log/traces` directory.
Create a handler for the home page with links to traces and resources.

Test Plan:
Ran `buckd` and went to http://localhost:9011/.
Verified that I saw the expected HTML.