Use a specific file to represent a successful build of buck.
Summary:
Previously, we were getting cases where dx or the nailgun client would
fail to build.  Buck would fail, but the next run of buck would see
buck.jar present, and try to build, causing weird failures down the
road.

This diff replaces buck.jar (which was just being used as a flag,
really) with an explicit successful-build file, which is created at the
end of the build.

Test Plan:
Injected a fail in to the dx build.  Ran buck repeatedly, and saw that
it never actually executed the buck java code.