Use `Paths` in `HasAndroidResourceDeps`.
Summary:
Making this change makes it easier to address the TODO in Buildable
about making `getPathToOutputFile` return a `Path` without causing too
many changes in the code base. In the process, a change was also made to
`BuildRuleFactoryParams`, and this was rippled through the handful of
non-Android files that were impacted by this.

Test Plan: `buck test --all`