Adding compiler option to buckconfig.
Summary:
I've added a compiler section to the `.buckconfig`. This is supposed to
be used to specify a custom compiler. I was thinking about wrapping this
in a class that does some additional checking whether the provided
command actually exsits but I'm not sure if that's worth the effort.

Test Plan: `buck test --all`