Add `BuildRuleParams.copyWithChangedDeps`.
Summary:
Previously, in order to make a copy of a `BuildRuleParams` with a
different set of deps, we had to invoke the normal constructor and pass
the right value to each parameter.  This is risky, because if someone
created a new constructor with an extra argument and made the old
constructor delegate to the new one with some default value, then these
"copies" would now have the default value for that parameter, rather
than copying it.

This introduces `copyWithChangedDeps`, which lives right next to the
constructor, to perform this operation.  If someone adds a new
constructor now, hopefully they'll notice to update this method.

Test Plan:
RuleKey for fb4a doesn't change.