Clean up totalDeps handling in `AndroidBinaryGraphEnhancer`
Summary:
Previously we were passing `preDexDeps` to
`addBuildablesToCreateAaptResources` so it could add them to the
`totalDeps` returned from the `AndroidBinaryGraphEnhancer`.  I need to
eliminate that because I need to set up the dexing steps after the aapt
steps (so they can reference R.java).

Now, we maintain an `ImmutableSortedSet.Builder` for `totalDeps` and
allow both dexing and aapt to add to it.

Test Plan:
RuleKey for fb4a doesn't change.