Automatically incorporare source paths from `getInputsToCompareToOutput` in the rule key.
Summary:
"Fix rulekey of AndroidResourceRule" fixed a bug in how we generate rule
keys for android_resource rule, essentially including source paths as
well as the source file contents in the rule key. Turns out `java_library`
rule also has the same bug.

This diff makes it so that we automatically include the source paths for
inputs returned from `Buildable#getInputsToCompareToOutput()` in the
rule key.

Should we restrict the scope of the `Path`/`SourcePath` setter methods on
`RuleKey.Builder` to package-private?

Test Plan: Unit tests.