Group `TestResults` together based on dependencies.
Summary:
`TestResultsGrouper#post()` does several things:

* Builds a graph of depencies between `TestResults` by following `deps`
  and `source_under_test` edges,
* Groups `TestResults` so that a parent test result is only posted if
  its dependencies have been posted as well, and
* Sets a flag saying whether a test result is reliable or not, based on
  whether it has failing dependencies.

For example, consider the following dependencies:

```
  A -(deps)-> {B}
  ATest -(source_under_test)-> {A}
  BTest -(source_under_test)-> {B}
```

Which is turned into this graph of test dependencies:

```
  ATest -> {BTest}
  BTest -> {}
```

When `BTest` is posted, it is immediately returned because it has no
dependencies to wait on, but if `ATest` is posted, it will only be
returned unless the dependencies of the library it is testing have also
been tested (which in this case is by `BTest`.)  If it isn't ready to be
returned, it is queued, and when `BTest` is posted then the clique of
`ATest` and `BTest` is returned.  If `BTest` failed, `ATest` has
`hasDependenciesPassTheirTests` set to `false`.

Test Plan: Mock test included.