New class `SourceSigner` to update and validate generated source file signatures.
Summary:
For generated IDE files (IntelliJ, Xcode, etc.) which need to be checked
in, we want to put a unique marker inside that contains a checksum of
the file.

If the developer accidentally changes the generated file, we can
catch the change via lint or a remote commit hook.

Test Plan: Unit tests included.