Add buck test selectors.
Summary:
Select which tests to run with a `--test-selectors` option.  A
`--filters` alias is provided for Facebook, who use this confusing
legacy name in other tools!

Then test selectors are given, test result caching is disabled and no
cached results persist.

Test result cache behavior is affected by this diff.
`TestCommand#isTestRunRequiredForTest()` is forced to always return true
if we are running with test-selectors.  Prior to this diff, a test would
not necessarily be run, in particular if result files from a previous
run were present then the test would not be re-run and the results would
be re-read from file (ie: "caching".)

Additionally, this diff means the result files that are written by the
`JUnitRunner` have an extra string in the file name (`.test_selectors`)
to stop runs done with test-selectors from interfering with
non-test-selectors runs.

```
  buck-out/gen/test/com/example/__java_test_util_output__/com.example.Foo.xml
  buck-out/gen/test/com/example/__java_test_util_output__/com.example.Foo.test_selectors.xml
```

(In buck, these files are dual purpose, being used for both the
results-cache **and** the means of communication between the child java
process using `JUnitRunner` and the parent java process that is buck.)

`JavaTestRule#hasTestResultFiles()` has not been modified to account for
differing result files names.  If we ever want to change buck to use
cached results when using test-selectors, JavaTestRule will have to be
updated either (a) to look at the cached-result file that has the
`.test_selectors` component or (b) we should cache results between
using-test-selectors and not-using-test-selectors runs.

Test Plan:
Unit and integration tests included!

In particular
`TestSelectorsIntegrationTest#shouldNeverCacheWhenUsingSelectors()`
exists to show that results from tests run with test selectors are never
used as cached results when running without selectors, and vice versa.
(If you run "test --all" followed by "test --all --filter X" followed by
"test --all" no cached results will be used at all.)