Add Xcconfig loader and evaluator.
Summary:
In the interest of faster transitioning from xcode to buck, instead of
replicating options in the xcode build system to equivalent options in
the buck rules, simply use the existing xcodebuild configuration
formats.

Background on xcconfig format and evaluation:
https://our.intern.facebook.com/intern/dex/fbobjc-tour/xcode-build-configuration/

Xcode configurations are implemented in a series of "layers" (instead of
executed imperatively), with inheritance between layers. The input
configuration stack has 4 layers. This diff will evaluate the 4 layers
to create a single layer of configuration, which is suitable for
embedding in an xcode target. (the input is derived from project +
target configs, which allows for 4 layers, while the output is only
target-level configs, which only allows for two layers).

The net effect is that if the user changes the original xcconfig files,
the project must be re-generated (which I think is sensible).

Test Plan:
```
ant clean default compile-tests
buck test test/com/facebook/buck/apple/xcode/xcconfig
```