Require `.buckconfig` to be present before running Buck.
Summary:
This is a common complaint: people run buck from the wrong directory,
and it leaves a buck-out dir that they have to manually clean up.  Let's
require a `.buckconfig` to be present before we execute the Buck Java
code.

I updated the documentation for quickstart, but some existing projects
might need to add a `.buckconfig` file when they pull in this change.
(Gerrit already has one, though.)

Test Plan:
- Ran `buck build fb4a` in fbandroid.  Saw buck start building.
- Ran `buck build fb4a` in fbandroid/java.  Saw an error message.
- Ran `buck build fb4a` in an empty directory.  Saw the error.
- Ran `touch .buckconfig && buck quickstart` in an empty directory.  Saw
  a new project created (with a proper .buckconfig).