Add type coercer `Either` to express heterogeneous types.
Summary:
Allows expressing that a field may consist of two different types (i.e. a
string or a list).

This is useful for some cases in the ios rules, where a field can be a file
name with some data, or a dictionary of the data written inline in the rule.
This should reduce the amount of either manual parsing, or massaging the result
in python using a wrapper rule.

Along with this I made the type traversals simpler. Type traversals visits all
the types in a type coercer. This was used to determine whether there's
something that looks like a build rule, in order to build a list of
dependencies (so that they can be loaded) before formally parsing the rule
itself (where it will require the loaded dependencies). I've replaced it with a
query function for whether the type coercer has some particular types.

Test Plan: Unit tests.