Bucket pre dex files into primary and secondary dexes earlier.
Summary:
We don't have to wait for the previous steps to complete bucketing pre
dex inputs into primary and secondary dexes. This was not possible
earlier because `AndroidBinaryRule` was responsible for dexing `R.java`,
which now is done by `UberRDotJava` which being a dependency is
guaranteed to finish before `AndroidBinaryRule` is built.

A subsequent diff will get rid of suppliers as well to simplify the code.

Test Plan:
`buck build messenger`
Edit a java file and verified only one of the secondary dex files is re-built.

Also planning to add a unit test for the Sorter class in another diff.