Drop persisted abi key from `DexProduced`.
Summary:
This is written, but not read in `initializeFromDisk`.  The comment is
not even accurate, because the ABI that we look at in
`DexProducedFromJavaLibraryThatContainsClassFiles` is not the Java ABI,
but rather the full Java class hashes.

Test Plan:
Unit tests are reasonably comprehensive.  Adding a character to a
comment in a java file in fbandroid forces a re-javac but not a re-dex.