Do not apply buck test timeouts when running with --debug.
Test Plan:
Create .buckconfig.local with:
```
[test]
    timeout = 10000 # 10 seconds
```

With and without this fix, run
`buck test --debug //javatests/com/facebook/resources/impl:impl`, setup a
breakpoint in one of the tests, attach the debugger and do not resume
execution for at least 10 seconds. Verify that there is no timeout error
with this fix.