Add missing components to apkbuilder command line
Summary:
These were being added programmatically, but not reflected in the
description.

Test Plan:
Ran the apkbuilder command displayed when building FB4A.  Got an
effectively-identical apk from the one produced by Buck. It was
unsigned, because ApkBuilderMain doesn't have arguments for the
keystore, and the files were in a different order in the zip file and
the signature.