Reorder some code in AndroidBinaryRule
Summary:
I wanted the code for the native libraries to be closer to the code for
apkbuilder, because they are both needed for the ABI key in exopackages.

I am currently wondering if getBuildSteps and getAbiKeyForDeps should be
unified into one method that returns both, because it is so important
for them to agree.

Test Plan: Buildbot.