Make AndroidBinaryRule an AbiRule (for exopackage)
Summary:
For pre-dexed builds, AndroidBinaryRule just runs ApkBuilder (plus a few
cp commands to set up its directory structure, and zipalign), and for
exopackage builds, that apkbuidler command doesn't include secondary dex
files.  Therefore, we can skip the entire rule if only secondary dex
files have changed.

For non-exopackage, we just make the abi key equal to our RuleKey, which
should force a rebuild every time the RuleKey doesn't match.

For exopackage builds, it's very important for the ABI key to reflect
all of the information that goes into ApkBuilder.  We start with the
saved SHA-1 hashes of the ap_ and primary dex.  We then manually hash
all of the native libraries, third party JARs, and keystore.  These are
small enough in FB4A that the hasing is pretty fast.  We can always
pre-hash them later.

Test Plan:
Integration tests.

I don't currently have a good way of ensuring that the ABI key and steps
to execute stay in sync.  I am currently wondering if getBuildSteps and
getAbiKeyForDeps should be unified into one method that returns both,
because it is so important for them to agree.