Add dx code from KitKat.
Summary:
This is an export of the dx directory from the dalvik repository from
0f45d4c669a4b771ecbf2dcb533394ed62ec8cd7, which is just what
origin/master is at today.

This is not MOE-protected because we are going to be including our
changes in the Open Source Buck release.

Test Plan:
Buildbot.
Verified that the git tree has the same hash.