Complete API methods in `ProjectFilesystem` for operating on `Paths`.
Summary:
Forward all calls with `String`s and `File`s to versions using `Path`s.

Plan is to slowly move all the code to using the new methods.

Not actually marking old versions of the methods as `@deprecated`
because it will trigger build warnings which we then turn to errors.

Test Plan: Tests pass.