Stable GID generation for targets.
Summary:
Allow xcode project object ids to be generated in a more stable manner.
Required for xcode->buck transition to allow build schemes specified in
xcode to refer to targets in the project (which is done through it's
gid).

The GID generated is based on a hash of the PBX object class, a hash of
the identifying data, and a counter for resolving ambiguities.
Individual classes can override GID generation behavior to use the
stable GID method, and pass it the relevant object specific hash.

Test Plan:
Added a unit test that verifies a target gets assigned a particular GID
derived from the method outlined above.