Use composition rather than inheritance for AdbCommandRunner
Summary:
AdbCommandRunner is providing functionality to the install and uninstall
commands by being a base class, but this is an antipattern.  What if
some other class wanted to use this functionality, but wasn't a
CommandRunner?  A better structure is to make AdbCommandRunner a service
object that the install and uninstall commands can call into.  This diff
makes that change.

Test Plan:
Unit tests
Installed and uninstalled an app from my emulator.