Installer for exopackage APKs
Summary:
The major components here are (1) an agent that gets installed on the
device to help with exopackage installation, (2) ExopackageInstaller
that runs in Buck to install the packages, and (3) integration with
InstallCommand to do this automatically in buck install.

Test Plan:
"time buck install fb4a" returned in sub-8 seconds.  App started
successfully.  (I hacked fbandroid to make it an exopackage.)