Add daemon integration tests.
Summary:
This change adds integration tests which check that the daemon notices a
variety of file changes and its invalidates its caches correctly. As
well as testing for a number of failures that have been seen in the wild
these tests also surfaced a number of inconsistencies between the way
the `WatchmanWatcher` and `WatchServiceWatcher` handled relative and
absolute paths within `WatchEvents`. This change also fixes those
problems so that all Paths are treated consistently and adds comments
and tests for the expected Path transformations.

Test Plan: `buck test --all`