Respect label priority in `buck test --exclude x --include x+y`.
Summary:
Selecting test rules with label sets was a bit janky.  We had two
options that respected inclusion before exclusion, which meant we
couldn't include everything with `labelA` but not if it had `labelB`.
The practical consequnce of this is in the attached task, and tested in
`TestCommandTest#testLabelPriority()`.

This diff also de-stringly-types labels in the codebase.

Test Plan: Included tests.