Set namespace aware parser for test results.
Summary:
We added some additional logging for this weird corner case where the
underlying xml parser seemingly gets out of sync and returns
obviously-bogus values for attributes.

A recent Q&A thread gave us some nice repros. I beat those (the xml)
over the head for a bit to isolate the problem. Net result: I have no
idea why this happens -- seems like a nonlocal parser problem because
the location of the failure jumps around when changing seemingly
unrelated parts of the file content.

On the bright side, at least for the two repros in ^ thread, the problem
goes away if we enable namespace parsing.