Use the optional custom javac if present to compile R.java files [part 1/2]
Summary:
Now that we allow using custom javac, we should use that everywhere in the code
base to compile any java code. This makes it so for the dummy R.java files that
are created per android library module.

The next part will do the same for the uber R.java files, which means
`AndroidBinaryRule` (which creates `UberRDotJava`) needs to be passed the path
to the custom javac.

Test Plan: sandcastle