Add proguard obfuscate/deobfuscate mapping function (part 3 of 4).
Summary:
This adds support for loading the ProGuard mapping file once, and using
that data to generate both an obfuscation function and a deobfuscation
function.

This is part of a series of changes to support scenario-guided
optimization of which classes to include in the primary dex.

The next change moves SplitDexStep over to use this since there are several
other changes to that file and it was simpler to divide the checkins this
way.