Refactor FilterResourcesStep to use Paths and ProjectFilesystem.
Summary:
I set out to write an integration test for resource filtering where I build
an android project, and inspect the APK programmatically to verify that it
contains the expected artifacts only. FilterResourcesStep is blocking that
effort since it uses `File.toPath()` a bunch of places instead of accessing
files through the project file system which handles relative/absolute paths
correctly.

Also, `Iterable<Path>` is bad because `Path implements Iterable<Path>`. This
uses `Collection<Path>` instead.

Someday, I'd like to refactor the class and move the scaling/drawable finder
jonx to separate files, but this is good enough for now.