Variant of ProjectFilesystem.walkFileTree() which works in integration tests.
Summary:
I was working on an integration test where `ProjectFilesystem.walkFileTree()`
would fail essentially saying 'root dir does not exist'. I was confused because
running the same buck build command as the integration test from inside both
the `.../testdata/project` or `/tmp/junitxxx` (where `ProjectWorkspace` sets
up a testdata project) would not fail!

So the reason the integration test was failing was because the current working
directory which running those would be `~/devtools/buck` and not `/tmp/junitxxx`.

This diff adds an explicit `ProjectFilesystem.walkRelativeFileTree()` and makes
sure that it works correctly in integration tests as well. I will send out
another diff that uses this function in FilterResourcesStep.