Advise buckd restart on error.
Summary:
Currently buckd exceptions are most likely caused by memory leaks
causing heartbeat timeouts to be missed due to garbage collection or
virtual memory thrashing. Until these leaks are fixed the simplest
workaround is to restart buckd, so this change advises that course of
action when an exception is thrown. In the cases where exceptions are
thrown for other reasons restarting buckd to get back to a known good
state is still a sensible course of action.

Test Plan:
1) `buckd`
2) `while true; do rm -r buck-out; rm -r buck-cache; buck build buck; done`
3) Check that exception message contains "if you are using buckd, you
   should restart it".