Include file change event processing time in parse time.
Summary:
Previously the time spent processing file change events was not included
in the parse time reported by buck, which overstated the benefits of
using buckd. This change records the time at the start of file change
event processing and uses that as the parse start time to more fairly
and accurately reflect the benefits of using buckd.