Add option to reference existing xcconfig in generated project.
Summary:
In separated projects mode, it's desirable to reference existing
xcconfig files directly instead of baking their contents into a single
config file, as this removes the need to re-generate the project when
the config files are changed and better integrates the generated file
with the existing xcode-centric build system.

This option only makes sense for configs which can be encoded in xcode's
project-target config structure (that is, a project level config file,
project level inline settings, target level config file, target level
inline settings).  This is why this is an option that should be turned
off for combined-mode projects.

This diff adds this functionality to map suitably laid-out config layers
into xcode's structure, or throw an error if such a mapping is not
possible.

Test Plan: Unit test for both positive and negative cases.