Allow specifying per-header visibility in BUCK files.
Summary:
Currently, we assume all headers are "public" visibility when generating
Xcode projects, which means they're visible to dependencies.

However, the default for headers is actually "project" visibility
(visible only to other targets in the project), and we didn't support
"private" visibility (visible only to the project(s) they're added to).

In this diff, I update the "headers" field of iOS library, binary, and
test targets to be closer to the "srcs" field. In particular, an entry
in "headers" can now either be:

1) A path to a header
2) A tuple of (header path, visibility flag)
3) A {group name, [headers...]} pair

Note: The iOS library, binary, and test targets now have quite a lot of
copy-pasta in their constructors. Any thoughts on how to best refactor
them?

Test Plan: Unit test updated.