Implement a build_config generator.
Summary:
This change set adds a build_config generator that knows how to
generate the BuildConfig.java files generated by existing build tools.
This addresses issue: https://github.com/facebook/buck/issues/3
and is based on the strategy proposed in the comment on that issue.

Test Plan:
There are already unit tests in place which verify all but the
contents of the generated file. To verify contents add a build_config
rule to a project and dig into buck-out to verify the contents
of the file are correct.