Compile uber R.java parallel with the aapt_package step.
Summary:
This diff moves the step that compiles non-english strings as assets to a new
buildable that also stores the hash of its output zip file.

As a result, aapt_package step no longer needs to wait for uber R.java to be
compiled and so those two can be executed in parallel.

Test Plan:
Updated unit tests.
Integration test for strings-as-assets passes.
AndroidBinaryRuleIntegrationTest passes.

Trace after changing a color attribute in one of the resource files in fbandroid:
https://www.facebook.com/pxlcld/l28j