Do not allow buck-out as an excluded folder in the root .iml file.
Summary:
Excluding all of `buck-out` causes problems for IntelliJ because it
looks in `buck-out/gen` for certain things (presumably autoproviders)
and gets upset when it cannot find them. By "gets upset," I mean it
creates this additional "modules" folder in the root of the file tree in
the left navigation pane, which makes it harder to navigate in IntelliJ.

Test Plan:
Ran `buck project` in fbandroid and verified that I saw the following
in `fbandroid_repo.iml`, as expected:

```
      <excludeFolder url="file://$MODULE_DIR$/buck-cache" />
      <excludeFolder url="file://$MODULE_DIR$/buck-out/bin" />
      <excludeFolder url="file://$MODULE_DIR$/buck-out/log" />
```

Note that the root `buck-out` directory is no longer in this list.

Most importantly, when I open IntelliJ, the annoying "modules" folder is
gone.