Support scenario-driven primary dex packing (part 4)
Summary:
This enables properties in android_binary() to allow specifying a list
of classes generated based on some scenario, such that those classes
and their first-order dependencies are included in the primary dex.

This is part of a series of changes to support scenario-guided
optimization of which classes to include in the primary dex.

Test Plan:
existing unit tests, build fb4a and messenger with startup classes list

Step through new code in debugger while building fb4a.