buck project with no arguments generates projects for all xcode_project_config rules
Summary:
Also change output to be relative paths instead of absolute, as is the
convention in intellij (also makes testing easier).

Test Plan:
Added integration test with 2 projects, and when run asserts that the list of projects
contains both projects. Did not check the contents or existence of actual files.