better handling of build setting prefixed framework paths
Summary:
* Adds the ability to prefix frameworks with a known build setting (currently
  only SOURCE_ROOT, SDKROOT, and BUILT_PRODUCTS_DIR).
* will now error when generating a project that makes use of a framework
  path that is prefixed by an unknown build setting.

note that this only catches the error when it is used (in test and binary
rules) and not when it is declared (in library rules). Doing that would require
the rule themselves be privy to SourceTreePaths, which will likely require some
file moving.

Note also that the build setting is evaluated in the context of the dependent
rule, not the dependency library in which its declared (since they are simply
copied verbatim into the output project). This means that these values,
especially those that are unprefixed or prefixed with SOURCE_ROOT, can produce
non-intuitive results.

Test Plan: unit test