Make ZipStep use Paths instead of Strings and Files.
Summary:
No one uses this step as of now, but I'll need it in D1202870. Since the
`baseDir` is not available when the zip step is created, I couldn't use this
step as is. Also `Path`s are better than strings and files.

`paths` argument only accepts files to include in the zip, not directories.

Test Plan: updated unit test passes.