Read target GIDs from existing Xcode project
Summary:
This builds on diff D1231075 to read in an existing Xcode
project if present and re-use the target GIDs inside when generating
a new project.

This keeps GIDs stable when first converting to Buck, so we don't
have to touch all the Xcode workspaces in the repo, which contain
references to target GIDs.

Depends On D1231075

Test Plan:
Unit tests included. Tested by hacking up Xcode project
file with fake GIDs, re-ran buck project and confirmed GIDs stayed
the same instead of being replaced.