Log javac version to scuba
Summary:
This diff logs the javac version to scuba.  The main part of this
change is moving the lookup of the javac version inot Main.java
and then threading it's value through the various places it's
needed, including the scuba logger.

Test Plan:
Ran "ant clean test".  Also, currently setting up fbandroid build
environment to test with an actual build.