Allow custom proguard versions specified in .buckconfig.
Summary:
Add a new .buckconfig entry tools/proguard which allows the user to
specify a relative path from the project root to a version of the
proguard.jar to be used during the Android obfuscation step.

Test Plan:
Modified fbandroid to use proguard 4.11 and verified with buck build
--verbose 10 that the version of proguard was upgraded (proguard
produces output including its version each time it is run).

See D1244542.