Alter the signature of Description to allow subclassing to work cleanly.
Summary:
Given:

  public Arg implements ConstructorArg {
  }

  public Arg2 extends Arg {
  }

The current definition of Description doesn't allow Arg2 to be passed to a
class that extends `Description<Arg>`. That subclass can use covariant return
types to return a subclass of `Arg`, but it can't be passed back to
`createBuildable`. Work around this by some generics-fu.

Test Plan: buck test --all