Coerce missing collection parameter values to empty collections when marshalling args.
Summary:
When converting from Python to Java, we promise to coerce optional collection args on
`ConstructorArg` instances to an empty collection in order to make handling them
easier. This is only true if that field is present in the `BuildRuleParams` passed
to the argument marshaller. In production, this is always the case, but in tests we
take short cuts, and don't always Do The Right Thing.

Rather than modify all the existing tests, we'll make the `ConstructorArgMarshaller`
handle this case correctly.

Test Plan: buck test --all