Support for disguised native executables in android_binary
Summary:
I want the exopackage install agent to contain a native executable.
This is a big pain because it has to be built for every architecture and
packaged into the apk, then we need code on the device to extract the
right one into world-readable storage.

The good news is that we already have support for most of this.
ndk_library can build executables if they are specified in Android.mk,
and the Android installer can automatically extract the right one.

The bad news is that this only works for files named "lib*.so".

The semi-hacky solution is to automatically rename any native library
output called "*-exe" to "lib*.so" to ensure that ApkBuilder will
package it and the installer will unpack it.  It is then up to the app
to know how to execute it.

Test Plan: Built and ran the (upcoming) native exopackage installer.