Buffer contents of wanted-in-primary-dex items
Summary:
We do a pass on classes initially to add items that are required to be
in the primary dex.  At the same time, we remember the items we find
that would be good to put in the primary dex, but that are not
required.  After the initial pass, we go back and add as many of the
items from the wanted list.

However, by the time the second pass occurs, the files from the first
pass are closed.  This crashes when we try to use the entries we cached.

This change explicitly buffers the files we need for the second pass.

Test Plan: build katana and verify we hit this path in buck.