Get rid of IntermediateDexRule and move the AbiRule interface of the buildable.
Summary:
I also renamed the buildable class because it was getting really hard to keep
the code well formatted. Feel free to suggest a better name or to just revert it.

I want this change because I am working on another diff that gets rid of the
builder classes in buildables.

Test Plan: buck test --all