SmartDexingStep no longer relies on the crc of zip entries to hash input jars.
Summary:
The hashes of inputs to be dexed are passed from above into SmartDexingStep,
which it uses to decide whether a dex needs to be recreated. Since earlier, it
was using the zip entry crc for this, it also used to check for timestamps on
the hash output files (arghhh) which basically meant smart dexing never worked
on the first run post pulling-everything-from-cache. Now everything is awesome!

Test Plan:
buck build fb4a, messenger
change a java file, build again, verify only one of the dex files is recreated.
unit tests.