Add JavaFileParser, using the Eclipse ASTParser.
Summary:
The autodeps feature needs to be able to parse Java files. Previously we tried
to use the builtin parser in tools.jar, but that got us into a dependency
tangle that probably would've broken on some version of the JDK. It's cleaner
to just check in a parser, and Eclipse's seems like a good candidate.

Test Plan: New unit test included.