Expose JavacOptions from DefaultJavaLibrary.
Summary:
To parse Java files, we need to be able to read the JavaCompilerEnvironment of
a JavaLibrary. This is stored on JavacOptions, so we add a method to expose
that.

This is a little bit dangerous, because the javacOptions member is actually
mutable, and we reset it for AndroidLibrary rules. This isn't trivial to
refactor away, because the Android parts need a classpath that isn't available
when the BUCK files are being parsed. It's not an immediate problem for us,
because nothing is changing the jcEnv member, but it *could* be a problem in
the future if we're not careful.

Test Plan: Just a new accessor method.