Avoid using exceptions for flow of control.
Summary:
In Java, it's considered poor form to use exceptions to control the flow of control.
We used them briefly here since it was a nice way to get started quickly, but we can
determine at coercion time what the correct type is, so there's no need to use the
exception-based approach.

Test Plan: buck test --all