Improve how BuildRuleSourcePaths are packaged into JAR files when used as resources.
Summary:
Previously, if a generated file from another directory was included as a resource of a `java_library()`, it
was placed in the directory corresponding to the BUCK file that defined the `java_library()`.
This diff changes the heuristics of `CopyResourcesStep` so that a generated file used as a resource
will be placed in a directory that corresponds to the Java package in which it "originates," if appropriate.

For example, if the `.buckconfig` declares the following:

```
[java]
  src_roots = /java/
```

and `//java/com/example/res:gen_resource` is a `genrule()` that produces 'foo.txt',
and `//java/com/example:example` is a `java_library()` that declares
`//java/com/example:gen_resource` as a resource, then the generated JAR file
will contain the entry `com/example/res/foo.txt`. Previous to this diff, it would
appear as `com/example/foo.txt`.

Test Plan:
Created an integration test: `CopyResourcesStepIntegrationTest`.
Verified that `CopyResourcesStepIntegrationTest` failed without my changes.
The existing `CopyResourcesStepTest` passes without any changes.
Sandcastle builds.