Prioritize API-version build-tools over android-version build-tools.
Summary:
As far as I can tell, the only way to get a build-tools dir named "android-*"
is to have it bundled with the SDK you download. Updates installed by the SDK
package manager seem to have the "19.0.0" style name. Since those should
presumably be newer, we should prefer them over the Android version names.

https://github.com/facebook/buck/issues/112

Test Plan:
Modified the corresponding unit test. Put an "android-4.3" directory in my SDK.
Observe that this breaks the build before this diff, but not anymore.