don't do an eager filesystem crawl to create the build file tree
Summary:
BuildFileTree crawled the filesystem to build an in-memory
representation of the build file tree. On large filesystems, this takes
a long time.  This splits BuildFileTree into two implementation:
InMemoryBuildFileTree, which uses an in-memory representation and never
crawls the filesystem, and FilesystemBackedBuildFileTree, which scans
the filesystem as-needed.

Test Plan:
Added some tests to verify correct behavior of the new
FilesystemBackedBuildFileTree. The tests pass. buck build buck works.
Running buck on large directory hierarchies now goes much faster.