Shutdown cassandra client context after each command.
Summary:
Previously buck would wait for in flight cassandra requests
to complete after each command, but not shutdown the client, resulting
in connections being left open to the cassandra caches when multiple
commands are run using buckd. This change shuts down the client in
addition to waiting for requests to complete, to avoid leaking
connections to the cache.

Test Plan:
1) buck test --all
2) buckd
3) buck build buck
4) lsof -nP -p `cat .buckd/buckd.pid` | grep TCP
5) check that 4 lists no connections to cache hosts