Fix broken test by closing ProjectBuildFileParser properly.
Summary:
D910431 created a test method called
`onlyUseThisWhenTestingToFindAllTransitiveDependencies` that created a
ProjectBuildFileParser without closing it. For some reason, D1243758 caused
this dangling parser to start writing to stderr when it closed (presumably
because the process it was piping to was already dead). Wrapping the parser in
a try-with-resources fixes the problem.

Test Plan:
The failure only triggered on a clean test, and only when all tests
were running at once. Again, I'm not sure why.

`buck clean && buck test --all`