Eliminate EMPTY_VISITOR static in DX
Summary:
BytecodeArray uses EMPTY_VISITOR to avoid allocating a new BaseVisitor
for each call to parseInstruction with visitor == null.  However,
BaseVisitor has visible mutable state, so running this in multiple
threads could produce incorrect results.  (It's possible that this is
already happening in multi-threaded DX).  To simplify analysis, just
delete it and allocate a new one each time.  We'll have to trust the JVM
to make it a cheap allocation and GC (hopefully it's short-lived).

Test Plan: ant dx-tests