GenAidl should always specify an output file.
Summary:
There is never a case where gen_aidl doesn't generate an output file. Without parsing the file itself,
we don't know the name, and we don't want to force people to adopt a naming convention. As a result,
the output file is a src-zip for use with javac.

Test Plan:
buck test --all
Also roll buck back to master, build the prod apps, roll forward to this diff and build again after a clean.